rag_news_pipeline/requirements.txt
requests
beautifulsoup4
nltk
sentence-transformers
faiss-cpu
transformers
torch
streamlit